requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.0
ultralytics==8.3.170
opencv-python==4.12.0.88
numpy==2.2.6
pandas==2.3.2
Pillow==11.3.0
pytesseract==0.3.13
torch==2.8.0
torchvision==0.23.0
matplotlib==3.10.5
python-multipart==0.0.20
pydantic==2.11.7
PyYAML==6.0.2
